refactor: reorganize str-win into separate platform-specific modules
- Create separate windows.rs and unix.rs modules for better organization
- Implement cross-platform compatible string conversion functions
- Unix implementation provides equivalent functionality using standard Rust APIs
- All tests pass on both platforms
